+++ example/lib/app.dart
@@ -1,5 +1,6 @@
-import 'package:draftmode_notifier_example/screen/home.dart';
 import 'package:flutter/cupertino.dart';
+//
+import 'screen/home.dart';
 
 class App extends StatelessWidget {
   final GlobalKey<NavigatorState> navigatorKey;