+++ example/lib/entity/geofence.dart
@@ -1,0 +1,7 @@
+class GeofenceConfig {
+  final bool sendNotificationOnEnter = true;
+  final bool sendNotificationOnExit = true;
+  final bool notifyDialogOnEnter = true;
+  final bool notifyDialogOnExit = true;
+  GeofenceConfig();
+}