+++ example/lib/geofence/mode.dart
@@ -1,0 +1,1 @@
+enum DraftModeGeofenceMode { enter, exit}